[bench] fix average fps precision + make `draw canvas` colorful

+++ src/org/zeroxlab/benchmark/CaseCanvas.java
@@ -45,7 +45,7 @@ public class CaseCanvas extends Case{
 	    total  += fps;
 	}
 
-	result += "Average: fps = " + (total/length) + "\n";
+	result += "Average: fps = " + ((float)total/length) + "\n";
 	return result;
     }
 }
