+++ src/org/zeroxlab/benchmark/CaseGLCube.java
@@ -46,7 +46,7 @@ public class CaseGLCube extends Case{
 	    total  += fps;
 	}
 
-	result += "Average: fps = " + (total/length) + "\n";
+	result += "Average: fps = " + ((float)total/length) + "\n";
 	return result;
     }
 }
